Move Node.js app reload into script
If we fire SIGHUP vs node before it has a chance to register the
signal handler, then it will immediately exit with an error code 7.
By moving the reload into an external script we can first make sure
that the process is listening which should only be the case after we
have the signal handler loaded.